server/requirements.txt
# Database
sqlalchemy==2.0.48
psycopg2-binary==2.9.11

# API & Web Server
fastapi==0.135.1
uvicorn==0.42.0

# Tools & Scrapers
scrapling==0.4.2
curl-cffi==0.7.3
playwright==1.50.0
browserforge==1.2.4
pyyaml==6.0.3
pydantic==2.12.5
alembic==1.15.1

# Background Processing
celery[redis]==5.6.0
flower==2.0.1
redis==5.2.1
requests==2.32.3

# Testing
pytest==8.3.4
httpx==0.28.1